pyproject.toml
[project]
name = "autolongvideogeneration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chardet>=5.2.0",
    "faiss-cpu>=1.12.0",
    "google-genai>=1.47.0",
    "langchain-google-genai>=0.1.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "moviepy>=2.2.1",
    "openai>=1.95.0",
    "opencv-python",
    "scenedetect[opencv]>=0.6.7.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]



[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
